requirements.txt
setuptools
Pillow==10.1.0
python-docx==1.0.1
python-pptx==0.6.21
PyPDF2==3.0.1
moviepy==1.0.3
pydub==0.25.1
reportlab==4.0.4
img2pdf==0.4.4
svglib==1.5.1
python-magic==0.4.27
pypdf==3.17.1
odfpy==1.4.1
ebooklib==0.18
markdown==3.5.1
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
pydot==1.4.2
graphviz==0.20.1
wand==0.6.11
ffmpeg-python==0.2.0
pyyaml
groq
flask
python-dotenv
speechrecognition
pyaudio
numpy<2
scipy
scikit-learn
webrtcvad
pydub
flask-cors
flask-socketio
requests
edge-tts
pygame
langdetect
flask>=2.3.2
pymongo>=4.6.3
bcrypt>=4.0.1
pycryptodome>=3.10.1
python-dotenv>=0.19.0
google-auth>=2.3.3
google-auth-oauthlib>=0.4.6
flask-login
flask-admin==1.6.1
WTForms==3.0.1
